build(mingw): make the cross build warning-clean under the Linux strict flag set
The MinGW cross build already used the same strict flags as Linux (-Wall -Wextra
-Wpedantic -Wconversion -Wshadow -fanalyzer ...); this makes it actually build
clean under them — 0 warnings, and it passes -DBVNR_WERROR=ON like Linux CI.

The 78 warnings fell into three groups:

Format (61): GCC checked every printf/fprintf/snprintf %zu/%lld against its
ms_printf builtin archetype (msvcrt-flavoured), not the header's gnu_printf
attribute, so C99 conversions warned even though they print correctly at runtime.
Fixed globally for MinGW with __USE_MINGW_ANSI_STDIO=1 (selects MinGW's ANSI
stdio) + -fno-builtin-{printf,fprintf,snprintf} (lets the gnu_printf attribute
drive -Wformat). Applied via add_compile_*() so the fuzz harnesses, which use
their own warning set, are covered too.

Float-conversion (7): all from MinGW's <math.h> isfinite/isinf/signbit, which
are macros with an internal double->float cast (glibc uses true builtins, so
Linux is clean). The cast is in the system macro, not our code, so -Wconversion
keeps its integer checks and only -Wno-float-conversion is dropped (in
BVNR_C_FLAGS, after -Wconversion).

Conversion (LLP64, 8) and unused (2): real, fixed at the source.
  * bvn_port.h: read()/write() take size_t but MinGW's CRT count is unsigned int
    — wrap them (matching the MSVC _read/_write signature) so the bounded count
    converts explicitly.
  * bvn_float.c: a few int64_t/uint32_t -> long assignments (long is 32-bit on
    Windows) at sites whose values are bounded (prec <= 32768, exp within +-1e9),
    cast explicitly; identical result on LP64.
  * bvnr_conformance.c: tap_skip/evlog_equal are used only by the (POSIX-only,
    already #ifndef _WIN32) IUT runner — guard their definitions too.

Verified: mingw build 0 warnings and passes -Werror; ctest 78/78 under Wine;
Linux -Werror build still clean and float/conformance/datetime tests pass.

+++ src/utils/bvn_float.c
@@ -1181,8 +1181,12 @@ static int32_t bvnf_to_str_dec(const bvn_float_t *f, char *buf, size_t bufsize)
 {
 	if (f->_prec > BVN_FLOAT_MAX_PREC) return -1;
 	if (f->_exp  >  1000000000L || f->_exp < -1000000000L) return -1;
-	long prec = f->_prec;
-	long exp2 = f->_exp;
+	/* Explicit casts: long is 64-bit on LP64 (Linux) but 32-bit on Windows
+	 * (LLP64). _prec (<= BVN_FLOAT_MAX_PREC) and _exp (guarded to +-1e9 above)
+	 * both fit a 32-bit long, so the value is identical on every target; the
+	 * casts only document the intended (bounded) narrowing for -Wconversion. */
+	long prec = (long)f->_prec;
+	long exp2 = (long)f->_exp;
 	long e2   = exp2 - prec;
 	long E10;
 	{
@@ -1360,7 +1364,7 @@ bool bvn_float_from_double(bvn_float_t *f, double v)
 	f->_sign = neg ? -1 : 1;
 	f->_exp  = exp2;
 	memset(f->_d, 0, (size_t)f->_nlimbs * sizeof(bvn_limb_t));
-	long prec = f->_prec;
+	long prec = (long)f->_prec;   /* bounded; see bvnf_to_str_dec cast note */
 	int shift = (int)prec - 53;
 #if BVN_LIMB_BITS == 64
 	if (shift >= 0) {
@@ -1487,7 +1491,7 @@ static void bvnf_to_ieee_bin_direct(const bvn_float_t *f,
 	for (int i = 0; i < bits32; i++) bits[i] = 0;
 	if (f->_sign < 0)
 		bits[(total - 1) / 32] |= 1u << ((total - 1) % 32);
-	long be   = f->_exp - 1L + (long)bias;
+	long be   = (long)f->_exp - 1L + (long)bias;
 	long drop = (long)f->_prec - ((long)man_bits + 1L);
 	if (be <= 0) { drop += (1L - be); be = 0; }
 	uint32_t shint = bvnf_hint_limbs(f->_prec > (man_bits + 1u) ? f->_prec : (man_bits + 1u));
